fix: corrigido erros de build
