+++ frontend/src/app/(auth-routes)/Pacientes/page.tsx
@@ -1,5 +1,4 @@
 import BreadCrumb from '@/src/components/BreadCrumb';
-
 import PacientesHeader from '@/src/components/Headers/Headerpacientes';
 import NavBar from '@/src/components/NavBar';
 import TabelaPacientes from '@/src/components/Tabelas/TabelaPacientes';
