+++ frontend/src/app/(auth-routes)/layout.tsx
@@ -1,7 +1,5 @@
 import { getServerSession } from 'next-auth';
-
 import { redirect } from 'next/navigation';
-
 import { ReactNode } from 'react';
 
 // import { MenuComponent } from '../../components/NavBar/MenuComponente';
