+++ frontend/src/app/layout.tsx
@@ -1,10 +1,8 @@
 import type { Metadata } from 'next';
-
 import { Geist, Geist_Mono } from 'next/font/google';
 
 import './globals.css';
 
-import React from 'react';
 import ThemeRegistry from '../theme/ThemeRegistry';
 import Providers from './providers';
