+++ frontend/src/components/Tabelas/TabelaNovoAtendimento/index.tsx
@@ -1,4 +1,5 @@
 'use client';
+import AddIcon from '@mui/icons-material/Add';
 import {
   Box,
   Button,
@@ -15,10 +16,9 @@ import {
 
 const pacientes = Array.from({ length: 7 }, () => ({
   nome: `Nome Paciente`,
-  cpf: '123.456.789-10',
-  sus: 46743674,
-  nomeMae: 'Maria José da Silva',
-  entrada: '09:00'
+  nomeMae: 'Nome da mãe',
+  idade: '36 anos e 2 meses',
+  entrada: '13:00'
 }));
 
 export default function TabelaNovoAtendimento() {
@@ -30,9 +30,8 @@ export default function TabelaNovoAtendimento() {
             <TableHead>
               <TableRow sx={{ backgroundColor: '#e0ecff' }}>
                 <TableCell sx={{ paddingY: 1 }}>PACIENTE</TableCell>
-                <TableCell sx={{ paddingY: 1 }}>CPF</TableCell>
-                <TableCell sx={{ paddingY: 1 }}>SUS</TableCell>
                 <TableCell sx={{ paddingY: 1 }}>NOME DA MÃE</TableCell>
+                <TableCell sx={{ paddingY: 1 }}>IDADE</TableCell>
                 <TableCell sx={{ paddingY: 1 }}>HORÁRIO DE ENTRADA</TableCell>
                 <TableCell align="right" sx={{ paddingY: 1 }} />
               </TableRow>
@@ -46,9 +45,8 @@ export default function TabelaNovoAtendimento() {
                       {paciente.nome}
                     </Typography>
                   </TableCell>
-                  <TableCell sx={{ paddingY: 1 }}>{paciente.cpf}</TableCell>
-                  <TableCell sx={{ paddingY: 1 }}>{paciente.sus}</TableCell>
                   <TableCell sx={{ paddingY: 1 }}>{paciente.nomeMae}</TableCell>
+                  <TableCell sx={{ paddingY: 1 }}>{paciente.idade}</TableCell>
                   <TableCell sx={{ paddingY: 1 }}>{paciente.entrada}</TableCell>
                   <TableCell align="right" sx={{ paddingY: 1 }}>
                     <Button
@@ -66,7 +64,7 @@ export default function TabelaNovoAtendimento() {
                         }
                       }}
                     >
-                      INICIAR ATENDIMENTO
+                      Ver prontuário
                     </Button>
                   </TableCell>
                 </TableRow>
@@ -78,6 +76,27 @@ export default function TabelaNovoAtendimento() {
         {/* Paginação agora dentro da largura da tabela */}
         <Box sx={{ display: 'flex', justifyContent: 'center', mt: 2 }}>
           <Pagination count={5} page={1} color="primary" />
+        </Box>
+        <Box sx={{ display: 'flex', justifyContent: 'flex-end', my: 2 }}>
+          <Button
+            variant="outlined"
+            startIcon={<AddIcon />}
+            sx={{
+              borderColor: '#1351B4',
+              color: '#1351B4',
+              borderRadius: '50px', // Deixei bem arredondado como no print
+              textTransform: 'none',
+              fontWeight: 'bold',
+              paddingX: 3,
+              paddingY: 0.8,
+              '&:hover': {
+                backgroundColor: 'rgba(19, 81, 180, 0.04)',
+                borderColor: '#1351B4'
+              }
+            }}
+          >
+            Novo Paciente
+          </Button>
         </Box>
       </Box>
     </Box>
